+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [0.8.0] 2022-06-29
 ### Added
-- `dbconsole` rake/rails command (also aliased as `db`) for easy access to the DB console. Only PostgreSQL supported at this time.
+
+- `rails dbconsole` command support for Sequel (also aliased as `rails db`) for easy access to the DB console. See Readme for installation instructions.
 
 ## [0.7.0] 2022-06-24
 ### Added